Move chord data to separate directory
